Heart added + SW done(?)
first attempt on heart with arcs was no good after all, so it is now two
bezier curves then a straight line, then it closes itself.

After manually looking through the debug stuff, I think I have FINALLY
sorted out the serviceworker events, so that it says when there is an
update downloaded, and also when an update has happened.

Later, I can add a What's New bit on that 'toast' popup!