the path to sound #2
Successfully implemented the sound API, loading the audio sprite file,
buffering it, connecting it to a gainNode, and then to a destination. It
plays the entire sound file as expected.

Next step is to look at performance - can I play loads of bits of the
single file, or would it be better to create loads more audio vars with
their buffers (source) as a single segment of the large audio file?