the path to sound #3
successfully managed to convert the audiobuffer into a float32array,
then make a new audiobuffer, and put a slice of that array into the new
buffer's buffer! the audio sprite is born!!!!!

TODO: make a new audioBuffer up for each word in the sprite, and create
mechanism for filling the buffers, playing, and resetting them.
Hopefully there will be an event fired when the audioBuffer has ended!!!
Otherwise I'll have to resort to animationFrame, and checking whether
the duration of the audioBuffer has happened yet.