the path to settings #Done!
I also updated my volume button to be more child-friendly, by making it
colored, with the button smoothly changing color as it is slid along
(Well-proud of it!!!!) - the percent isn't needed, and I put the volume
'icon' inside the button.

Also added asking to save preferences to the toast popup, and updated
that a bit too.